don't cast between long and pointers for win64
long is 32 bit there

+++ frontend/main.c
@@ -998,7 +998,7 @@ void *SysLoadLibrary(const char *lib) {
 		tmp++;
 		for (i = 0; i < ARRAY_SIZE(builtin_plugins); i++)
 			if (strcmp(tmp, builtin_plugins[i]) == 0)
-				return (void *)(long)(PLUGIN_DL_BASE + builtin_plugin_ids[i]);
+				return (void *)(uintptr_t)(PLUGIN_DL_BASE + builtin_plugin_ids[i]);
 	}
 
 #if !defined(_WIN32) && !defined(NO_DYLIB)
@@ -1013,7 +1013,7 @@ void *SysLoadLibrary(const char *lib) {
 }
 
 void *SysLoadSym(void *lib, const char *sym) {
-	unsigned int plugid = (unsigned int)(long)lib;
+	unsigned int plugid = (unsigned int)(uintptr_t)lib;
 
 	if (PLUGIN_DL_BASE <= plugid && plugid < PLUGIN_DL_BASE + ARRAY_SIZE(builtin_plugins))
 		return plugin_link(plugid - PLUGIN_DL_BASE, sym);
@@ -1036,7 +1036,7 @@ const char *SysLibError() {
 }
 
 void SysCloseLibrary(void *lib) {
-	unsigned int plugid = (unsigned int)(long)lib;
+	unsigned int plugid = (unsigned int)(uintptr_t)lib;
 
 	if (PLUGIN_DL_BASE <= plugid && plugid < PLUGIN_DL_BASE + ARRAY_SIZE(builtin_plugins))
 		return;
